Added eventrewrited and fixed Jetway remote
eventrewrited is a supercharged version of
eventlircd. It allow you to map multiple key
combinations into a single key and allows you
to remap your buttons any way you want.

This release contains a fix for the Jetway Remote.
All the buttons how work properly and are mapped to
the correct functions. Due to not having a button
to display the context menu the "Magnifier" key has
been doubled into two keys. If you press it, it will
act as the Info key. If you press and hold it for
atleast 0.3 seconds it will popup the context menu.

One step closer to the perfect HTPC disto :)