Add a simple URL checker
scripts/checkurl is a simple URL checker which checks the validity of
package source file URLs without downloading the actual files.

When run with no parameters, it scans through all package meta files and
checks the URLs specified in PKG_URL variables.
When run with a package name as parameter, it checks URLs for that package.

If VERBOSE=yes is specified in options, it will output the entire server
response and parsed reason result, which is useful for finding out why a
particular URL doesn't work or for debugging.

Signed-off-by: Alain Kalker <a.c.kalker@gmail.com>