Easily deploy your new OpenELEC image
If the DEPLOY_TO_OPENELEC_APPLIANCE variable is defined in the
environment, (it should contain the IP/hostname of your OpenELEC
appliance), at the end of the building process the newly built image
would be uploaded automatically to it over SCP.

The new image files and their MD5 checksums will be put in the special
software update directory inside the /storage filesystem, and the
appliance will be rebooted in order to run the new image.

It is recommended that you configure key-based SSH login to the
appliance to make this process easier.

If the variable is not defined a message would be shown, telling the
user about the existence of this possibility.