Skip internal mounts for EFI and Recovery
On AppleTV devices both EFI and Recovery are shown in the XBMC interface as a source for media, however both partitions are not required for OpenELEC during XBMC runtime, so we avoid mounting them by default. As a positive side-effect it also prevents triggering hfsplus and vfat modules during boot.

Note that once we go back to single-boot mode, Recovery will be mounted during boot, so this change should not affect any progress in that field.

I also cleaned up the script so it needs less invocations of commands, and less system utilities are forked for parsing output (as we use awk now). So it should be somewhat faster.