Add fstab for easy mounting of EFI and Recovery partitions
Since we disabled automatic mounting of EFI and Recovery on boot (to avoid having them listed as disks containing media) we may want to allow both to be mounted without much effort by adding entries in fstab. For completeness we also add the other filesystems/swap. The missing part is provisioning the (empty) mount points, but there's no infrastructure to create directories just for projects/ATV.