Make tvheadend not send debug message to syslog
Apparently tvheadend by default uses the syslog service, and the `-s` option that we provide in the startup script in fact makes tvheadend send debug messages to syslog (which we obviously do not want by default as this causes for my single DVB card 1k/min to /var, which is 4MB/month). So this fix disables the `-s` option in the tvheadend startup script.