Fix slow syslogd remote syslog startup and klogd interaction
Because syslogd is not properly started its remote syslog support when klogd dumps its kernel ring buffer content to it, most of the kernel boot output is missing on the remote syslog server. By adding a small delay between syslogd and klogd, we are sure the messages are properly send to the remote syslog by syslogd.