avahi: samba service cosmetics
When samba is built with avahi support the server name is advertised in
UPPERCASE as samba forces this when setting-up avahi via dbus. If we
disable avahi support in samba and include a samba.service file with
avahi the hostname is correctly advertised as the user entered it; e.g.
"OpenELEC" not "OPENELEC" ..we can also use a tweaked samba.service so
Mac users with an AppleTV see the AppleTV icon in Finder.app instead of
a generic server icon.