Added patch for sql_lex.h
This enables us to compile MySQL with server support (i.e. remove `--without-server` `--without-embedded-server`) in `package.mk`.

This is useful for people wanting to use the MySQL daemon on the Raspberry Pi for example.

However, this commit should be seen as a discussion topic, since the environment variable `YACC_HEXT_HH` seems to be set, but it doesn't build the .hh file (builds the regular .h file) so this patch pretty much works around that by hardcoding the sql_yacc.h-reference instead.