add package for MT7601U wireless chipsets
This resolves #2524 and requests in the forums. It builds but I have no
hardware to test so do not merge unless there are reports of success.
nb: the MT7601USTA.dat file is not a true firmware (it’s config) but I
figured /lib/firmware is a tidier place to put it than the default
/etc/Wireless location.