Add package keyutils
The kernel hardcodes the path to the key instantiation callback binary
as /sbin/request-key, meaning that providing keyutils as a separate
binary in an addon is impossible. It might be possible to use a
dangling symlink, but the three binaries built from this package
are all around 20 - 30KiB. Additionally, the request-key binary
is useful for more than one protocol.

keyutils is the "new" way of handling both DNS name resolution
(necessary for e.g. cifs dfs referrals) and security credentials, thus
replacing e.g. the rpc.idmap daemon.

Signed-off-by: David Härdeman <david@hardeman.nu>