Enable all buttons on the TiVo Slide Pro remote.
The linux kernel has supported the TiVo Slide remote control for some time,
but does not recognize the USB ID of the newer Slide Pro. This patch adds
the missing data structures so the newer remote will be recognized by the
driver, thereby allowing the TiVo, LiveTV, and Thumbs Up/Down buttons to be
mapped with a hwdb file.