fuse : fix unavailable package due to bad url
fuse packages are not available on sourceforge so url:
PKG_URL="$SOURCEFORGE_SRC/fuse/fuse-2.X/$PKG_VERSION/$PKG_NAME-$PKG_VERSION.tar.gz"
is deprecated and building the image with:
 PROJECT=imx6 ARCH=arm make image
fail because archive cannot be found.

Now the repo is available on github:
https://github.com/libfuse/libfuse.git

This patch modify the makefile informations for fuse and add a script to generate the package.

Before merging this patch, the archive genrated with ./tools/mkpkg/mkpkg_fuse must be copied on "$DISTRO_SRC/" mirror.
Be careful that the commit id has not changed since the creation of this patch.