Enable Getty & Login in Busybox for UART Login
Many dev boards (in particular RPi and RPi2) have a hardware UART interface. Currently the only way to get a shell in Lakka is via ssh. In some situations, it is not possible/desirable to setup a network to ssh in to the RPi/2. Enabling Getty in busybox allows a shell to be spawned on the hardware UART interface. 

Note: A user will still have to place `/sbin/getty -L ttyAMA0 115200 vt100` in either `autostart.sh` or self-compile as a system.d service.