Fix bug in Noritake VFD driver
Hello,
there is a bug in the driver NoritakeVFD.c, see patch. Columns and rows are swapped for setting the cursor position.
It took me a lot of time to find this error and get the display working.
Hope it helps others.
Tested with Noritake CU20045SCPB-T31A connected via serial line to RPi2 as well as desktop with Ubuntu.
Regards
P.S. The Noritake VFD displays are really nice to build a media player with "old school touch". For reasonable prices you can get it on the well known online market auction platform.