Force udev to detect Shield Controller as Controller
The shield controller has a touchpad at the bottom of it. udev/evdev by default takes this and assumes that the shield controller is not a controller, but a mouse/touchpad. This rule forces udev to present a shield controller as a controller, and not a joypad. /dev/js* (linuxraw) inputs are broken without this rule too.