Add --disable-bindings to llvm configure.
This fixes building for me on Ubuntu 14.04. Otherwise, llvm can't find
ocaml ctypes in order to build llvm with ocaml bindings.

The ocaml-ctypes package isn't available in Ubuntu 14.04. So, it's easier
to just leave it out because it doesn't seem like Lakka needs ocaml
bindings for llvm for anything else.