RPI2: Add SYSFS to disable RJ45 leds.
This patch adds the following sysfs entries which allow control of RJ45 LEDs. 0 turns the LED off, 1 turns the LED on (default 1)
/sys/class/smsc95xx_leds/eth_fdx
/sys/class/smsc95xx_leds/eth_lnk
/sys/class/smsc95xx_leds/eth_spd

This commit was found here: https://www.anginf.de/?p=433 and I am only submitting as a commit for a Pull Request so that this can be incorporated into official builds. By adding this, it enables RPI2 users to completely disable all lights from the RPI2. Prior to this patch the RJ45 LEDs flashing could be annoying depending on the location of the RPI2.